Greatly simplified the artist pages. Used mui grid to make it compatible with mobile
